backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.12
PyJWT==2.9.0
reportlab==4.2.2
openpyxl==3.1.5
